+++ examples/StateComponents/transitionOnEnter.ts
@@ -1,8 +1,8 @@
 import { StateSetupFn } from '../..';
-import { IMachineSPI } from '../..';
+import { ISetupMachine } from '../..';
 
 export const transitionOnEnter = <T extends string>(
     state: T
 ): StateSetupFn<T, any> => {
-    return (machine: IMachineSPI<T, any>) => machine.transitionToState(state);
+    return (machine: ISetupMachine<T, any>) => machine.transitionToState(state);
 }