+++ src/IAuxDataHoldingMachine.ts
@@ -1,4 +1,4 @@
-export interface IAuxDataContainer<D> {
+export interface IAuxDataHoldingMachine<D> {
     getAuxillaryData(): D | undefined;
     setAuxillaryData(data: D): void;
 }