+++ src/IMachine.ts
@@ -1,4 +1,4 @@
-import { IAuxDataContainer } from './IAuxDataContainer';
+import { IAuxDataHoldingMachine } from './IAuxDataHoldingMachine';
 import { MachineBlueprint } from './types';
 
 /**
@@ -6,7 +6,7 @@ import { MachineBlueprint } from './types';
  * presented to state components (e.g., when states are entered, run, or
  * exited), see IMachineSPI.
  */
-export interface IMachine<D> extends IAuxDataContainer<D> {
+export interface IMachine<D> extends IAuxDataHoldingMachine<D> {
     init<S extends string>(blueprint: MachineBlueprint<S, D>, auxillaryData?: D): void;
     isInitialized(): boolean;
     terminate(): void;