+++ src/ISetupMachine.ts
@@ -1,0 +1,10 @@
+import { IAuxDataHoldingMachine } from './IAuxDataHoldingMachine';
+
+/**
+ * Internal interface for TinyFSM state machines. This interface is injected
+ * into state component setup functions.
+ */
+export interface ISetupMachine<T extends string, D> extends IAuxDataHoldingMachine<D> {
+    transitionToState(state: T | 'end'): void;
+    terminate(): void;
+}